+++ RELEASE_NOTES.md
@@ -1,3 +1,3 @@
-### Breaking Change
+### New Features
 
-API classes have been refactored to all be under the `Api/` namespace, with the exception of `Client`, which has not moved.
+- Implement Subscribers API